# docs/swagger.yaml
basePath: /
definitions:
...
  models.PreferredLanguageRequest:
    properties:
      language_code:
        example: en
        type: string
    required:
    - language_code
    type: object
...
  models.User:
    properties:
      email:
        type: string
      full_name:
        description: Otomatik oluşturulan tam ad
        type: string
      id:
        type: string
      name:
        description: Kullanıcının adı
        type: string
      password:
        type: string
      phone_number:
        type: string
      preferred_language:
        description: Kullanıcı tercihi
        type: string
      roles:
        description: '["admin", "editor", "user"]'
        items:
          type: string
        type: array
      surname:
        description: Kullanıcının soyadı
        type: string
      username:
        type: string
    required:
    - email
    - name
    - password
    - roles
    - surname
    - username
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is the API documentation for the admin panel project.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Admin Panel API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Retrieve all users with their details
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Failed to retrieve users
          schema:
            additionalProperties: true
            type: object
      summary: Get all users
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Add a new user with roles and hashed password
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create user
          schema:
            additionalProperties: true
            type: object
      summary: Create a new user
      tags:
      - Users
  /users/{id}:
    delete:
      description: Remove a user by its unique identifier
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid user ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete user
          schema:
            additionalProperties: true
            type: object
      summary: Delete a user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user details including name, roles, and password
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated user details
        in: body
        name: user
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload or user ID
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update user
          schema:
            additionalProperties: true
            type: object
      summary: Update a user
      tags:
      - Users
  /users/language:
    put:
      consumes:
      - application/json
      description: Update the preferred language of the current user
      parameters:
      - description: Preferred language request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.PreferredLanguageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Preferred language updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid language code or disabled language
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update preferred language
          schema:
            additionalProperties: true
            type: object
      summary: Update preferred language
      tags:
      - Users
security:
- BearerAuth: []
swagger: "2.0"